dags/my_simple_dbt_dag.py
from airflow.decorators import dag

from airflow.providers.postgres.operators.postgres import PostgresOperator
from cosmos import DbtDag, DbtTaskGroup, ProjectConfig, ProfileConfig, ExecutionConfig
from airflow.utils.dates import days_ago

default_args = {
    'owner': 'airflow',
    'start_date': days_ago(1),
}


