requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
#apache-airflow-providers-github==2.7.0
#astronomer-cosmos==1.0.4
#apache-airflow-providers-google==10.22.0
#astronomer-cosmos[dbt-bigquery]==1.0.3 // install google + cosmos + dbt
#protobuf==3.20.0
#REMOVE apache-airflow-providers-google==10.3.0
#REMOVE soda-core-bigquery==3.0.45
astronomer-cosmos[dbt-bigquery]==1.0.3
protobuf==3.20.0


